python: Improve child process termination
Previously we always killed the child process, even if it was not
running, and we did not log anything.

Now we first poll child process status and get its exit code. If the
child process did not terminate, we terminate it.

With this flow, we can log useful message when we terminate the child
process, and when the child process terminates.

Here are example flows:

Child process (pid=999116) was terminated by another process, and a new
child was started (pid=999118):

00:43:37,555 INFO    (MainThread) [IOProcessClient] (ioprocess-0) Starting client
00:43:37,560 INFO    (ioprocess/999116) [IOProcessClient] (ioprocess-0) Poll error 16 on fd 18
00:43:37,560 INFO    (ioprocess/999116) [IOProcessClient] (ioprocess-0) ioprocess was terminated by signal 15
00:43:37,565 INFO    (ioprocess/999118) [IOProcess] (ioprocess-0) Starting ioprocess

Child process (pid=999118) was closed normally:

00:43:37,570 INFO    (MainThread) [IOProcessClient] (ioprocess-0) Closing client
00:43:37,571 INFO    (ioprocess/999118) [IOProcessClient] (ioprocess-0) Shutdown requested
00:43:37,571 INFO    (ioprocess/999118) [IOProcessClient] (ioprocess-0) Killing ioprocess
00:43:37,572 INFO    (ioprocess/999118) [IOProcessClient] (ioprocess-0) ioprocess was terminated by signal 9

+++ bindings/python/ioprocess/__init__.py
@@ -179,12 +179,24 @@ def _communicate(ioproc_ref, proc, readPipe, writePipe):
         if (evtReciever >= 0):
             os.close(evtReciever)
 
-        if IOProcess._DEBUG_VALGRIND:
-            os.kill(proc.pid, signal.SIGTERM)
-        else:
-            proc.kill()
+        rc = proc.poll()
 
-        proc.wait()
+        if rc is None:
+            real_ioproc._log.info("(%s) Killing ioprocess", real_ioproc.name)
+            if IOProcess._DEBUG_VALGRIND:
+                os.kill(proc.pid, signal.SIGTERM)
+            else:
+                proc.kill()
+            rc = proc.wait()
+
+        if rc < 0:
+            real_ioproc._log.info(
+                "(%s) ioprocess was terminated by signal %s",
+                real_ioproc.name, -rc)
+        else:
+            real_ioproc._log.info(
+                "(%s) ioprocess terminated with code %s",
+                real_ioproc.name, rc)
 
         real_ioproc = ioproc_ref()
         if real_ioproc is not None:
